refactor(todisk): use source dirs relative to build dir

+++ src/generators/output/toDisk.js
@@ -3,8 +3,8 @@ const fs = require('fs-extra')
 const fm = require('front-matter')
 const glob = require('glob-promise')
 const deepmerge = require('deepmerge')
-const { asyncForEach } = require('../../utils/helpers')
 const removePlaintextTags = require('../../transformers/plaintext')
+const { asyncForEach, getPropValue } = require('../../utils/helpers')
 
 const Config = require('../config')
 const Tailwind = require('../tailwind')
@@ -16,27 +16,32 @@ module.exports = async (env, spinner) => {
   const config = await Config.getMerged(env).catch(err => { spinner.fail('Build failed'); console.log(err); process.exit(1) })
   const css = await Tailwind.fromFile(config, env).catch(err => { spinner.fail('Build failed'); console.log(err); process.exit(1) })
 
-  const sourceDir = config.build.posthtml.templates.root
-  const outputDir = config.build.destination.path
+  const sourceDir = getPropValue(config, 'build.templates.root') || 'src/templates'
+  const outputDir = getPropValue(config, 'build.destination.path') || `build_${env}`
+  let filetypes = getPropValue(config, 'build.templates.extensions') || 'html'
 
-  await fs.remove(outputDir)
-
-  if (Array.isArray(sourceDir)) {
-    await asyncForEach(sourceDir, path => fs.copy(path, outputDir))
-  } else {
-    await fs.copy(sourceDir, outputDir)
+  if (fs.pathExistsSync(outputDir)) {
+    await fs.remove(outputDir)
   }
-
-  let filetypes = config.build.posthtml.templates.extensions || 'html'
 
   if (Array.isArray(filetypes)) {
     filetypes = filetypes.join('|')
   }
 
+  if (Array.isArray(sourceDir)) {
+    await asyncForEach(sourceDir, async source => {
+      await fs.copy(source, outputDir).catch(error => spinner.warn(error.message))
+    })
+  } else {
+    await fs.copy(sourceDir, outputDir).catch(error => spinner.warn(error.message))
+  }
+
   const templates = await glob(`${outputDir}/**/*.+(${filetypes})`)
 
   if (templates.length < 1) {
-    spinner.fail(`Error: no files with the .${filetypes} extension found in ${path.resolve(outputDir)}`).fail('Build failed')
+    spinner
+      .fail(`Error: no files with the .${filetypes} extension found in your \`templates.root\` path${Array.isArray(sourceDir) ? 's' : ''}`)
+      .fail('Build failed')
     process.exit(1)
   }
 
@@ -78,8 +83,6 @@ module.exports = async (env, spinner) => {
       }
     }
 
-    const ext = templateConfig.build.destination.extension || 'html'
-
     if (templateConfig.plaintext) {
       await Plaintext.output(html, file, templateConfig)
     }
@@ -87,18 +90,24 @@ module.exports = async (env, spinner) => {
     html = removePlaintextTags(html, config)
 
     fs.outputFile(file, html)
-      .then(() => {
+      .then(async () => {
         if (templateConfig.permalink) {
-          return fs.move(file, templateConfig.permalink, { overwrite: true })
+          await fs.move(file, templateConfig.permalink, { overwrite: true })
         }
 
-        const parts = path.parse(file)
-        fs.rename(file, `${parts.dir}/${parts.name}.${ext}`)
+        const extension = getPropValue(templateConfig, 'build.destination.extension') || 'html'
+
+        if (extension !== 'html') {
+          const parts = path.parse(file)
+          await fs.rename(file, `${parts.dir}/${parts.name}.${extension}`)
+        }
       })
   })
 
-  if (fs.pathExistsSync(config.build.assets.source)) {
-    await fs.copy(config.build.assets.source, `${outputDir}/${config.build.assets.destination}`)
+  const assets = getPropValue(config, 'build.assets') || { source: '', destination: 'assets' }
+
+  if (fs.pathExistsSync(assets.source)) {
+    await fs.copy(assets.source, `${outputDir}/${assets.destination}`)
   }
 
   if (config.events && typeof config.events.afterBuild === 'function') {
